gates en controlador gastoController , en las rutas y vistas y en plantilla

+++ app/Http/Controllers/GastoController.php
@@ -4,56 +4,72 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Gasto;
+use Illuminate\Support\Facades\Gate;
 
 class GastoController extends Controller
 {
     public function ver(){
-        $gastos=Gasto::Paginate(15);
-        $monto =Gasto::sum('monto');
-        return view('gastos')->with('gastos',$gastos)->with('monto',$monto);
+        if(Gate::denies('isAdmin') || Gate::denies('isSecretaria')){
+            $gastos=Gasto::Paginate(15);
+            $monto =Gasto::sum('monto');
+            return view('gastos')->with('gastos',$gastos)->with('monto',$monto);
+        }else{
+            abort(403);
+        }
     }
     //  public function creargasto(){
     //     return view('nuevogasto');
     // } 
     
     public function guardargasto(Request $request){
-            
-        $request->validate([
-            'categoria'     =>  'required',
-            'detalle'       =>  'required',
-            'monto'         =>  'required|numeric|min:0|max:100000000000000000',
-            'fechafactura'  =>  'required|date',
-            'fechapago'     =>  'required|date',
-        ]);
+        if(Gate::denies('isAdmin') || Gate::denies('isSecretaria')){
+                    
+                $request->validate([
+                    'categoria'     =>  'required',
+                    'detalle'       =>  'required',
+                    'monto'         =>  'required|numeric|min:0|max:100000000000000000',
+                    'fechafactura'  =>  'required|date',
+                    'fechapago'     =>  'required|date',
+                ]);
 
-        // formulario
-        $nuevogasto = new Gasto();
-        $nuevogasto->categoria=    $request->input('categoria');
-        $nuevogasto->detalle=      $request->input('detalle');
-        $nuevogasto->monto=        $request->input('monto');
-        $nuevogasto->fechafactura= $request->input('fechafactura');
-        $nuevogasto->fechapago=    $request->input('fechapago');
+                // formulario
+                $nuevogasto = new Gasto();
+                $nuevogasto->categoria=    $request->input('categoria');
+                $nuevogasto->detalle=      $request->input('detalle');
+                $nuevogasto->monto=        $request->input('monto');
+                $nuevogasto->fechafactura= $request->input('fechafactura');
+                $nuevogasto->fechapago=    $request->input('fechapago');
 
-        $creado = $nuevogasto->save();
-        //Asegurarse que fue creado
-        if ($creado){
-            return redirect()->back()->with('mensaje','El nuevo Gasto fue creado exitosamente');
-        }else{ 
-        }
+                $creado = $nuevogasto->save();
+                //Asegurarse que fue creado
+                if ($creado){
+                    return redirect()->back()->with('mensaje','El nuevo Gasto fue creado exitosamente');
+                }else{ 
+                }
+            }else{
+                abort(403);
+            }
     }
     public function borrargasto($id){
-        Gasto::destroy($id);
-        return redirect()->back()->with('mensaje','Gasto borrado satisfactoriamente');
+        if(Gate::denies('isAdmin')){
+            Gasto::destroy($id);
+            return redirect()->back()->with('mensaje','Gasto borrado satisfactoriamente');
+        }else{
+            abort(403);
+        }
     }
 
     /* funcion para poder editar un gasto */
     public function editar($id){
-     
-        $gastos=Gasto::findOrFail($id);
-        return view('editargasto')->with('gastos',$gastos);
-
+        if(Gate::denies('isAdmin')){
+            $gastos=Gasto::findOrFail($id);
+            return view('editargasto')->with('gastos',$gastos);
+        }else{
+            abort(403);
+        }
     }
     public function update(Request $request,$id){
+        if(Gate::denies('isAdmin')){
         $request->validate([
             'categoria'     =>  'required',
             'detalle'       =>  'required',
@@ -76,6 +92,9 @@ class GastoController extends Controller
             return redirect()->back()->with('mensaje','¡¡El  Gasto Fué Modificado Exitosamente!!');
         }else{ 
         }
+    }else{
+        abort(403);
+    }
 
     }
